refactor: rename ui module to controller
